requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
psycopg2==2.8.6
python-dotenv==0.19.0
Werkzeug==2.0.1
bcrypt==4.0.1
Flask-Mail==0.9.1
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
